reference/real-time-ml-system-cohort-3/services/price-predictor/pyproject.toml
[project]
name = "price-predictor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "comet-ml>=3.47.5",
    "elasticsearch>=8.17.0",
    "joblib>=1.4.2",
    "loguru>=0.7.3",
    "optuna>=4.1.0",
    "pydantic-settings>=2.6.1",
    "quixstreams>=3.4.0",
    "hopsworks>=4.1.3",
    "scikit-learn>=1.6.0",
    "xgboost>=2.1.3",
    "pyarrow>=18.1.0",
]
